Submitted the hook rule which was failing
Message received from Uri: You can update your rule; it will be counted even though the project is officially closed because it took us a while to return to you

Thank you for reporting this. We will work on making that error message clearer, with the locations of the hooks.